Renamed BlackButton to PrimaryButton and adjusted code.
